ensure match of the *whole* prefix
... before removing the prefix from the mailbox name

The faulty behaviour can be tested as follows:

 * have a mailbox with the name foobar
 * make a store 'my_mails' in your config with a prefix 'foorious'

now if you launch:

 $ mailsync my_mails

the listing will include a mailbox with the name 'bar'
(that is 'foobar' minus the 'foo' from 'foorious')

Thanks a lot to Peter Breuer for finding the bug, doing the analysis
and proposing a patch! The patch was modified and applied by T.Pospíšek.